Add files via upload
Reimplement the CashRegister class so that it keeps track of the total price as an integer: the total cents of the price. For example, instead of storing 17.29, store the integer 1729. Such an implementation is commonly used because it avoids the accumulation of roundoff errors. Do not change the public interface of the class.

Kept vector implementation from 9.22.6, but readded total cost private data member.